requirements.txt
django==3.2.12
djangorestframework==3.13.1
django-extensions==3.1.5
django-filter==21.1
ipdb==0.13.9
pytest==6.2.2
pytest-django==4.5.2
